[VIBE] 2k25 AoC Rust Day 1
The first day of vibing with AntiGravity and Gemini 3 Pro (High)

Learned that it can solve problems such as setting up the enviorement.
Especially good with solving the fresh WSL distro networking issues if
VPN is enabled on the connecting machine; otherwise the host won't be
able to connect to the internet.

The first part it used rem_euclid() - remaining modulu function.
Since it solved for a problem I didn't even know would arrise of when
the modulu will return negative values for negative numbers.

Also for the second part it had a choice between simulating every single
click (which would be slow) and having a mathematical solution of which
was just checking how many times the modulu limit was reached by simple
division ((dial + amount) / 100).